Initial commit for SCAG (Scaffolding for Gramine)
With this commit, we unveil the fundamental elements of this,
mechanism: 'setup' (enabling users to create a scaffold project),
and 'build' (generating a Docker image for the user's application).

To assist new users, we have introduced a 'quickstart' command,
which allows users interactively configures their projects.

At present, we offer support for Flask and plain Python.

Co-authored-by: Wojtek Porczyk <woju@invisiblethingslab.com>
Co-authored-by: Mariusz Zaborski <oshogbo@invisiblethingslab.com>
Signed-off-by: Wojtek Porczyk <woju@invisiblethingslab.com>
Signed-off-by: Mariusz Zaborski <oshogbo@invisiblethingslab.com>